Beginning Collision Added state to paddles that tell you which direction the paddle is traveling in order to change the velocity of the ball correctly based on the area the ball hits on the paddle. Fixed checking for when you hold both the up and down key at the same time.
Further Steps:

Need to figure out how to change velocity of the ball. We are 
considering positive and negative numbers in how they relate to the 
speed. We are also considering upper and lower limits of the ball and 
that the ball shouldn't ever have a velocity of 0. We need to determine 
the maximum possible velocity. Could do (abs(current velocity of that 
axis) + amount the velocity increases )* sign to make it negative if it 
is negative.